add stimuli
