+++ allensdk/test/brain_observatory/behavior/conftest.py
@@ -72,8 +72,8 @@ from allensdk.brain_observatory.behavior.data_objects.metadata\
 from allensdk.brain_observatory.behavior.image_api import ImageApi
 from allensdk.brain_observatory.behavior.session_apis.data_transforms import \
     BehaviorOphysDataTransforms
-from allensdk.brain_observatory.behavior.stimulus_processing import \
-    StimulusTemplateFactory
+from allensdk.brain_observatory.behavior.data_objects.stimuli.\
+    stimulus_processing import StimulusTemplateFactory
 from allensdk.test_utilities.custom_comparators import WhitespaceStrippedString
